Trait num_digitize::FromIterRadix

pub trait FromIterRadix where Self: Copy + Sized + Num + NumCast {
    fn from_iter_radix<I: IntoIterator<Item=Self>>(iter: I, base: i64) -> i64 { ... }
}

Provided Methods

fn from_iter_radix<I: IntoIterator<Item=Self>>(iter: I, base: i64) -> i64

Converts a iterable of numbers back to original number of given base.

Example

Basic usage:

use num_digitize::FromIterRadix;

let original_number = 12345678;
let output = i8::from_iter_radix([1, 2, 3, 4, 5, 6, 7, 8].iter().cloned(), 10);
assert_eq!(original_number, output);

Please note, bases other than 10 were not tested and may not give any sense!